Firmware: Fix pots not going to 0

+++ firmware/src/core_m4/controls.hh
@@ -53,7 +53,7 @@ struct Controls {
 	void start();
 	void update_params();
 
-	uint32_t get_pot_reading(uint32_t pot_id);
+	float get_pot_reading(uint32_t pot_id);
 	uint32_t get_patchcv_reading();
 	uint32_t get_jacksense_reading();
